requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pymysql==1.1.0
python-dotenv==1.0.0
gtts==2.4.0
cryptography==41.0.0
